Fixes time_zone parameter in composite aggregation
